respond to PR comments

+++ apps/prairielearn/src/models/questions.sql
@@ -25,7 +25,7 @@ SELECT
     when q.type = 'Freeform' then 'v3'
     else 'v2 (' || q.type || ')'
   end AS display_type,
-  coalesce(issue_count.open_issue_count, 0) AS open_issue_count,
+  coalesce(issue_count.open_issue_count, 0)::int AS open_issue_count,
   row_to_json(top) AS topic,
   (
     SELECT
